+++ src/TestExplorer/Helpers/constants.ts
@@ -7,10 +7,12 @@ export default class Constants {
   static readonly individualTestPassedMessage: string = "OK";
   static readonly invalidTestMessage: string = "Test Failed: Check Terminal Output";
 
-  static readonly classOutputRegex = /(Tests:.*Assertions.*(Incomplete|Risky|Skipped|Failures)(?:\w*[^\r\n\.])*)/is;
+  static readonly classOutputRegex = /(Tests:.*Assertions.*(Incomplete|Risky|Skipped|Failures|Errors)(?:\w*[^\r\n\.])*)/is;
 
   static readonly phpUnitSuccessRegex = /OK\s+\(.*\)/i;
   static readonly phpUnitFailureRegex = /Failed\s+(?:[^\.])*/i;
+  // TODO: Determine the Regex to display the correct message on an error.
+  static readonly phpUnitErrorRegex = /Error Regex*/i;
   static readonly phpUnitNoTestsRegex = /No tests executed/i;
   static readonly phpUnitNoAssertionsRegex = /This test did not perform any assertions/i;