+++ test/tests/Unit/Unit with Spaces/PassingTestWithSpacesTest.php
@@ -13,11 +13,6 @@ class PassingTestWithSpacesTest extends TestCase
      */
     public function spacesTest()
     {
-        // Arrange
-
-        // Act
-
-        // Assert
         $this->assertTrue(true);
     }
 }